Acabados los mappers

+++ src/main/java/com/imo/cementery/model/dto/pago/PagoResponseDTO.java
@@ -6,6 +6,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
@@ -15,7 +16,7 @@ import java.time.LocalDate;
 public class PagoResponseDTO {
 
     private Long id;
-    private Double importe;
+    private BigDecimal importe;
     private LocalDate fecha;
     private PagoType metodo;
